Nodes render correctly and uniquely
Changed from creating each node independently to having an array keep track of which characters are already present. There's currently no reason for a character to be removed from this list that I can see so far, so by adding another character to this array and then re-rendering, we effectively create another node.