Basic saving and loading
Task 11 - Save the list availableCharacters as txt file, comma-separated
Task 12 - Import a txt file of the same structure. Read the file and update the character array to have only those characters.

These tasks are the main purpose of v0.3.

Also, changed availableCharacters to a useState hook. (Previously we only needed to append but importing a file overwrites instead.)